Move responsability to download images to the GenericComic.update() method
The get_next_comic method returns a dict with a 'url' attribute corresponding
to a (potentially empty) list of url of images. The update method will use it
to try to retrieve the images and set the 'local_img' attribute which will
corresponding to the images that were successfully downloaded.

In order to set the prefix to be use, a prefix attribute can be added but is
optional.

This makes things more consistent for comics with one image and comics with a
variable number of images.