Fix support for Garfield - #33
Garfield official website has changed : url are
formatted differently, images have different names
and are hosted on a different machine.
Anyway, the old code was meant to break at some
point as it relies on meant-to-be-broken
assumptions.

Rewriting Garfield as a NavigableComic is quite
easy and should be easier to fix in the future.

I have not fully tested the code as I've
re-triggered the download of ALL comics.