Handle GoComics with no nav on landing page
Issues that was reported to me by email:
the retrieval of first link does not quite work
on some GoComics such as Peanuts.
Indeed, for these special cases, the landing
page does not correspond to a page with the
navigation arrow. In order to find the nav
link, one has to follow the "Comics" link.

Note: this was not found with the navigation
testing function because a link was found
anyway: the issue is that it linked to the
last comics. Thus, this led to the retrieval
of one comic.

For the reference, here a the comics for which
is makes a difference (first: the url we use
to find, second: the url we now return):

From http://www.gocomics.com/calvinandhobbes
http://www.gocomics.com/comics/lists/1643203/calvin-and-hobbes-bus-stop-musings
http://www.gocomics.com/calvinandhobbes/1985/11/18

From http://www.gocomics.com/foxtrot
http://www.gocomics.com/comics/lists/1715369/foxtrot-new-years-resolutions
http://www.gocomics.com/foxtrot/1988/04/11

From http://www.gocomics.com/garfield
http://www.gocomics.com/comics/lists/1720856/garfield-loneliness-denial-weird-storyline-1989
http://www.gocomics.com/garfield/1978/06/19

From http://www.gocomics.com/lunarbaboon
http://www.gocomics.com/comics/lists/1722545/20-parenting-comics-lunarbaboon
http://www.gocomics.com/lunarbaboon/2015/10/05

From http://www.gocomics.com/peanuts
http://www.gocomics.com/comics/lists/1626936/peanuts-at-the-ballpark
http://www.gocomics.com/peanuts/1950/10/02

From http://www.gocomics.com/pearlsbeforeswine
http://www.gocomics.com/comics/lists/1721839/pearls-before-swine-writing-process-gets-meta
http://www.gocomics.com/pearlsbeforeswine/2002/01/07

From http://www.gocomics.com/sarahs-scribbles
http://www.gocomics.com/comics/lists/1720899/sarahs-scribbles-depression-comics
http://www.gocomics.com/sarahs-scribbles/2014/01/02

From http://www.gocomics.com/wumo
http://www.gocomics.com/comics/lists/1722213/thanksgiving-turkey-point-of-view
http://www.gocomics.com/wumo/2013/10/13

From http://www.gocomics.com/the-awkward-yeti
http://www.gocomics.com/comics/lists/1722479/best-heart-and-brain-comics-for-heart-month-awkward-yeti
http://www.gocomics.com/the-awkward-yeti/2014/09/08