Better support for dates
Since the implementation of dates with
"""
commit 507469b4a109d23f2b6983817cb391df511b211a
Author: Sylvain <sylvain.desodt+github@gmail.com>
Date:   Thu Aug 7 19:00:31 2014 +0100
"Add better support for dates by adding the dates in the abstract logic if needed"
"""
everything was done under the assumption that
the date in the dictionnary representing the
comic would be split in 3 integer fields: day,
month and year (because that's how things were
done for xkcd).

This would be easier to use if a proper 'date'
object was provided in the 'date' field and
then converted in the common logic.

This removes a lot of boilerplate code.
A lot more can be removed now that we do not
need the 'day' variables everywhere anymore.
(This will be done in a different commit)