Fix retrieval from tumblrs (photo -> Photo in values from API)
For some reason, it seems like now, the API returns "Photo"
as the type in places where "photo" used to be expected.

I do not know why and this does not seem to be aligned
with the documentation https://www.tumblr.com/docs/en/api/v1:

"type - The type of posts to return. If unspecified or empty,
all types of posts are returned. Must be one of text, quote,
photo, link, chat, video, or audio."

I've contacter the support, just in case.