Define a global setting to enable/disable checks
Performing checks can be useful from a few point
of view but can lead to additional requests to
the servers and can lead to misleading logs
to the users. Easiest solution is to have a
global setting to control it.

Also, I took this chance to rename/document
a method whose name started with "check_" but
is actually useful to an end user.